+++ frontend/src/views/Settings/Setting/Setting.vue
@@ -26,6 +26,7 @@
                     non-owners.
                 -->
                 <SettingScreenshotRetention />
+                <SettingAutoCloseProjects />
                 <SettingTimeReminder />
             </div>
         </div>
@@ -44,6 +45,7 @@ import SettingMilestoneStatus from "@/components/molecules/Setting/SettingMilest
 import settingFileExtesnsions from "@/components/molecules/Setting/SettingFileExtensions.vue";
 import SettingCurrency from "@/components/molecules/Setting/SettingCurrencys.vue";
 import SettingScreenshotRetention from "@/components/molecules/Setting/SettingScreenshotRetention.vue";
+import SettingAutoCloseProjects from "@/components/molecules/Setting/SettingAutoCloseProjects.vue";
 import SettingTimeReminder from "@/components/molecules/Setting/SettingTimeReminder.vue";
 import { defineComponent} from "vue";
 import { useCustomComposable } from '@/composable';
@@ -59,6 +61,7 @@ defineComponent({
         settingFileExtesnsions,
         SettingCurrency,
         SettingScreenshotRetention,
+        SettingAutoCloseProjects,
         SettingTimeReminder
     }
 })